+++ src/foundation/filter.rs
@@ -1,7 +1,9 @@
 #![allow(clippy::empty_docs)]
 
+use std::cmp::Ordering;
 use std::collections::HashMap;
 use enum_iterator::all;
+use indexmap::IndexSet;
 use log::{debug, error, log_enabled, trace, Level};
 use pest::iterators::Pair;
 use pest::Parser;
@@ -67,9 +69,16 @@ impl ValueProcessor for MockValueProcessor {
 }
 
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
+#[serde(untagged)]
+pub enum TemplateValue {
+    Single(String),
+    Multi(Vec<String>),
+}
+
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
 pub struct PatternTemplate {
     pub name: String,
-    pub value: String,
+    pub value: TemplateValue,
 }
 
 #[derive(Debug, Clone)]
@@ -254,7 +263,7 @@ fn get_parser_regexp(
         let mut parsed_text = String::from(expr.as_str());
         parsed_text.pop();
         parsed_text.remove(0);
-        let regstr = apply_templates_to_pattern(&parsed_text, templates);
+        let regstr = apply_templates_to_pattern_single(&parsed_text, templates)?;
         let re = regex::Regex::new(regstr.as_str());
         if re.is_err() {
             return create_tuliprox_error_result!(TuliproxErrorKind::Info, "cant parse regex: {}", regstr);
@@ -414,7 +423,7 @@ pub fn get_filter(
 ) -> Result<Filter, TuliproxError> {
     let empty_list = Vec::with_capacity(0);
     let template_list: &Vec<PatternTemplate> = templates.unwrap_or(&empty_list);
-    let source = apply_templates_to_pattern(filter_text, template_list);
+    let source = apply_templates_to_pattern_single(filter_text, template_list)?;
 
     match FilterParser::parse(Rule::main, &source) {
         Ok(pairs) => {
@@ -490,15 +499,21 @@ fn build_dependency_graph(
     let mut graph = DirectedGraph::<String>::new();
     for template in templates {
         graph.add_node(&template.name);
-        CONSTANTS.re_template_var
-            .captures_iter(&template.value)
-            .filter(|caps| caps.len() > 1)
-            .filter_map(|caps| caps.get(1))
-            .map(|caps| String::from(caps.as_str()))
-            .for_each(|e| {
-                graph.add_node(&e);
-                graph.add_edge(&template.name, &e);
-            });
+        let mut handle_template_value = |value| {
+            CONSTANTS.re_template_var
+                .captures_iter(value)
+                .filter(|caps| caps.len() > 1)
+                .filter_map(|caps| caps.get(1))
+                .map(|caps| String::from(caps.as_str()))
+                .for_each(|e| {
+                    graph.add_node(&e);
+                    graph.add_edge(&template.name, &e);
+                });
+        };
+        match &template.value {
+            TemplateValue::Single(value) => handle_template_value(value),
+            TemplateValue::Multi(values) => values.iter().for_each(|value| handle_template_value(value)),
+        }
     }
     let cycles = graph.find_cycles();
     for cyclic in &cycles {
@@ -520,7 +535,7 @@ pub fn prepare_templates(
     templates: &Vec<PatternTemplate>,
 ) -> Result<Vec<PatternTemplate>, TuliproxError> {
     let graph = build_dependency_graph(templates)?;
-    let mut template_values = HashMap::<String, String>::new();
+    let mut template_values = HashMap::<String, TemplateValue>::new();
     let mut template_map: HashMap<String, PatternTemplate> = templates
         .iter()
         .map(|item| {
@@ -533,11 +548,43 @@ pub fn prepare_templates(
         if let Some(sorted) = graph.topological_sort() {
             for template_name in sorted {
                 if let Some(depends_on) = dependencies.get(&template_name) {
-                    let mut templ_value = template_values.get(&template_name).unwrap().to_string();
+                    let mut templ_value = template_values.get(&template_name).unwrap().clone();
                     for dep_templ_name in depends_on {
-                        let dep_value = template_values.get(dep_templ_name).ok_or_else(|| TuliproxError::new(TuliproxErrorKind::Info, format!("Failed to load template {dep_templ_name}")))?;
-                        templ_value =
-                            templ_value.replace(format!("!{dep_templ_name}!").as_str(), dep_value);
+                        let dep_value = template_values.get(dep_templ_name).ok_or_else(|| info_err!(format!("Failed to load template {dep_templ_name}")))?;
+                        templ_value = match dep_value {
+                            TemplateValue::Single(dep_val) => {
+                                match templ_value {
+                                    TemplateValue::Single(templ_val) => TemplateValue::Single(templ_val.replace(format!("!{dep_templ_name}!").as_str(), dep_val)),
+                                    TemplateValue::Multi(templ_vals) => {
+                                        let mut new_values = vec![];
+                                        for val in templ_vals {
+                                            new_values.push(val.replace(format!("!{dep_templ_name}!").as_str(), dep_val));
+                                        }
+                                        TemplateValue::Multi(new_values)
+                                    }
+                                }
+                            }
+                            TemplateValue::Multi(dep_vals) => {
+                                match templ_value {
+                                    TemplateValue::Single(templ_val) => {
+                                        let mut new_values = vec![];
+                                        for dep_val in dep_vals {
+                                            new_values.push(templ_val.replace(format!("!{dep_templ_name}!").as_str(), dep_val));
+                                        }
+                                        TemplateValue::Multi(new_values)
+                                    }
+                                    TemplateValue::Multi(templ_vals) => {
+                                        let mut new_values = vec![];
+                                        for dep_val in dep_vals {
+                                            for templ_val in &templ_vals {
+                                                new_values.push(templ_val.replace(format!("!{dep_templ_name}!").as_str(), dep_val));
+                                            }
+                                        }
+                                        TemplateValue::Multi(new_values)
+                                    }
+                                }
+                            }
+                        };
                     }
                     template_values.insert(template_name.clone(), templ_value);
                 }
@@ -552,12 +599,80 @@ pub fn prepare_templates(
     Ok(template_map.into_values().collect())
 }
 
-pub fn apply_templates_to_pattern(pattern: &str, templates: &Vec<PatternTemplate>) -> String {
-    let mut new_pattern = pattern.to_string();
+pub fn apply_templates_to_pattern(
+    pattern: &str,
+    templates: &Vec<PatternTemplate>,
+    allow_multi: bool,
+) -> Result<TemplateValue, TuliproxError> {
+    let mut new_pattern = TemplateValue::Single(pattern.to_string());
+
     for template in templates {
-        new_pattern = new_pattern.replace(format!("!{}!", &template.name).as_str(), &template.value);
+        let placeholder = format!("!{}!", &template.name);
+
+        match &template.value {
+            TemplateValue::Single(val) => {
+                match new_pattern {
+                    TemplateValue::Single(ref mut pat) => {
+                        *pat = pat.replace(&placeholder, val);
+                    }
+                    TemplateValue::Multi(ref mut pats) => {
+                        for pat in pats.iter_mut() {
+                            *pat = pat.replace(&placeholder, val);
+                        }
+                    }
+                }
+            }
+
+            TemplateValue::Multi(ref multi_vals) => {
+                new_pattern = match &new_pattern {
+                    TemplateValue::Single(pat) => {
+                        let mut new_values = IndexSet::new();
+                        for val in multi_vals {
+                            new_values.insert(pat.replace(&placeholder, val));
+                        }
+                        TemplateValue::Multi(new_values.into_iter().collect())
+                    }
+                    TemplateValue::Multi(pats) => {
+                        let mut new_values = IndexSet::new();
+                        for val in multi_vals {
+                            for pat in pats {
+                                new_values.insert(pat.replace(&placeholder, val));
+                            }
+                        }
+                        TemplateValue::Multi(new_values.into_iter().collect())
+                    }
+                };
+            }
+        }
     }
-    new_pattern
+
+    if !allow_multi {
+        match &new_pattern {
+            TemplateValue::Single(_) => {}
+            TemplateValue::Multi(multi_vals) => {
+                match multi_vals.len().cmp(&1) {
+                    Ordering::Less => {
+                        return create_tuliprox_error_result!(TuliproxErrorKind::Info, "Empty multi value templates are not supported for pattern! {pattern}");
+                    }
+                    Ordering::Equal => {
+                        new_pattern = TemplateValue::Single(multi_vals.first().unwrap().to_owned());
+                    }
+                    Ordering::Greater => {
+                        return create_tuliprox_error_result!(TuliproxErrorKind::Info, "Multi value templates are not supported for pattern! {pattern}");
+                    }
+                }
+            }
+        }
+    }
+
+    Ok(new_pattern)
+}
+
+pub fn apply_templates_to_pattern_single(pattern: &str, templates: &Vec<PatternTemplate>) -> Result<String, TuliproxError> {
+    match apply_templates_to_pattern(pattern, templates, false)? {
+        TemplateValue::Single(value) => Ok(value),
+        TemplateValue::Multi(_) => create_tuliprox_error_result!(TuliproxErrorKind::Info, "Multi value templates are not supported for pattern!"),
+    }
 }
 
 #[cfg(test)]